Use an environment variable to change the name of DLLs
This change is specific to Windows and Windows CE builds,
but it can be ignored for all other builds.  Simply do not
define the environment variable named CRYPTO_PREFIX
unless you want to.

This is useful for defining a unique set of DLL names so that
multiple versions of OpenSSL can be present and yet an application
can select the version it requires.

Change-Id: I5af57ce08dd5f06d44d0f31d98d171619b5b333d